backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
pydantic
celery
redis
python-dotenv
ffmpeg-python
python-multipart



